requirements.txt
anthropic>=0.39.0
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0
